Fixed Soccer:Mode logic that determines which team(s) can score and which team is a goal's owner. Also, updated the interface to have freq # as a short.

+++ src/Core/ComponentInterfaces/IBalls.cs
@@ -200,6 +200,6 @@ namespace SS.Core.ComponentInterfaces
         /// </param>
         /// <exception cref="ArgumentNullException">Arena is null.</exception>
         /// <exception cref="Exception">Invalid <see cref="SoccerMode"/> for the arena.</exception>
-        void GetGoalInfo(Arena arena, int freq, MapCoordinate coordinate, out bool isScorable, out int? ownerFreq);
+        void GetGoalInfo(Arena arena, short freq, MapCoordinate coordinate, out bool isScorable, out short? ownerFreq);
     }
 }
